+++ core/src/db/DBImpl.h
@@ -27,9 +27,9 @@
 #include "config/handler/EngineConfigHandler.h"
 #include "db/DB.h"
 #include "db/IndexFailedChecker.h"
-#include "db/OngoingFileChecker.h"
 #include "db/Types.h"
 #include "db/insert/MemManager.h"
+#include "db/meta/FilesHolder.h"
 #include "utils/ThreadPool.h"
 #include "wal/WalManager.h"
 
@@ -183,20 +183,20 @@ class DBImpl : public DB, public server::CacheConfigHandler, public server::Engi
 
  private:
     Status
-    QueryAsync(const std::shared_ptr<server::Context>& context, const meta::SegmentsSchema& files, uint64_t k,
+    QueryAsync(const std::shared_ptr<server::Context>& context, meta::FilesHolder& files_holder, uint64_t k,
                const milvus::json& extra_params, const VectorsData& vectors, ResultIds& result_ids,
                ResultDistances& result_distances);
 
     Status
     HybridQueryAsync(const std::shared_ptr<server::Context>& context, const std::string& table_id,
-                     const meta::SegmentsSchema& files, context::HybridSearchContextPtr hybrid_search_context,
+                     meta::FilesHolder& files_holder, context::HybridSearchContextPtr hybrid_search_context,
                      query::GeneralQueryPtr general_query,
                      std::unordered_map<std::string, engine::meta::hybrid::DataType>& attr_type, uint64_t& nq,
                      ResultIds& result_ids, ResultDistances& result_distances);
 
     Status
     GetVectorsByIdHelper(const std::string& collection_id, const IDNumbers& id_array,
-                         std::vector<engine::VectorsData>& vectors, const meta::SegmentsSchema& files);
+                         std::vector<engine::VectorsData>& vectors, meta::FilesHolder& files_holder);
 
     void
     InternalFlush(const std::string& collection_id = "");
@@ -226,7 +226,7 @@ class DBImpl : public DB, public server::CacheConfigHandler, public server::Engi
     StartMergeTask();
 
     Status
-    MergeFiles(const std::string& collection_id, const meta::SegmentsSchema& files);
+    MergeFiles(const std::string& collection_id, meta::FilesHolder& files_holder);
 
     Status
     BackgroundMergeFiles(const std::string& collection_id);
@@ -235,7 +235,7 @@ class DBImpl : public DB, public server::CacheConfigHandler, public server::Engi
     BackgroundMerge(std::set<std::string> collection_ids);
 
     Status
-    MergeHybridFiles(const std::string& table_id, const meta::SegmentsSchema& files);
+    MergeHybridFiles(const std::string& table_id, meta::FilesHolder& files_holder);
 
     void
     StartBuildIndexTask();
@@ -254,10 +254,7 @@ class DBImpl : public DB, public server::CacheConfigHandler, public server::Engi
 
     Status
     GetFilesToBuildIndex(const std::string& collection_id, const std::vector<int>& file_types,
-                         meta::SegmentsSchema& files);
-
-    Status
-    GetFilesToSearch(const std::string& collection_id, meta::SegmentsSchema& files);
+                         meta::FilesHolder& files_holder);
 
     Status
     GetPartitionByTag(const std::string& collection_id, const std::string& partition_tag, std::string& partition_name);