+++ core/src/index/thirdparty/faiss/utils/distances.h
@@ -24,31 +24,6 @@ namespace faiss {
  * Optimized distance/norm/inner prod computations
  *********************************************************/
 
-#ifdef __AVX__
-/// Squared L2 distance between two vectors
-float fvec_L2sqr_avx (
-        const float * x,
-        const float * y,
-        size_t d);
-
-/// inner product
-float  fvec_inner_product_avx (
-        const float * x,
-        const float * y,
-        size_t d);
-
-/// L1 distance
-float fvec_L1_avx (
-        const float * x,
-        const float * y,
-        size_t d);
-
-float fvec_Linf_avx (
-        const float * x,
-        const float * y,
-        size_t d);
-#endif
-
 #ifdef __SSE__
 float fvec_L2sqr_sse (
         const float * x,