+++ core/src/server/Server.cpp
@@ -24,6 +24,7 @@
 #include "server/DBWrapper.h"
 #include "server/grpc_impl/GrpcServer.h"
 #include "server/init/CpuChecker.h"
+#include "server/init/GpuChecker.h"
 #include "server/web_impl/WebServer.h"
 #include "src/version.h"
 //#include "storage/s3/S3ClientWrapper.h"
@@ -189,7 +190,44 @@ Server::Start() {
         }
         tzset();
 
-        InitLog(log_config_file_);
+        {
+            bool trace_enable = false;
+            bool debug_enable = false;
+            bool info_enable = false;
+            bool warning_enable = false;
+            bool error_enable = false;
+            bool fatal_enable = false;
+            std::string logs_path;
+            s = config.GetLogsTraceEnable(trace_enable);
+            if (!s.ok()) {
+                return s;
+            }
+            s = config.GetLogsDebugEnable(debug_enable);
+            if (!s.ok()) {
+                return s;
+            }
+            s = config.GetLogsInfoEnable(info_enable);
+            if (!s.ok()) {
+                return s;
+            }
+            s = config.GetLogsWarningEnable(warning_enable);
+            if (!s.ok()) {
+                return s;
+            }
+            s = config.GetLogsErrorEnable(error_enable);
+            if (!s.ok()) {
+                return s;
+            }
+            s = config.GetLogsFatalEnable(fatal_enable);
+            if (!s.ok()) {
+                return s;
+            }
+            s = config.GetLogsPath(logs_path);
+            if (!s.ok()) {
+                return s;
+            }
+            InitLog(trace_enable, debug_enable, info_enable, warning_enable, error_enable, fatal_enable, logs_path);
+        }
 
         std::string deploy_mode;
         s = config.GetServerConfigDeployMode(deploy_mode);
@@ -204,25 +242,43 @@ Server::Start() {
                 return s;
             }
 
+            try {
+                // True if a new directory was created, otherwise false.
+                boost::filesystem::create_directories(db_path);
+            } catch (...) {
+                return Status(SERVER_UNEXPECTED_ERROR, "Cannot create db directory");
+            }
+
             s = InstanceLockCheck::Check(db_path);
             if (!s.ok()) {
                 std::cerr << "deploy_mode: " << deploy_mode << " instance lock db path failed." << std::endl;
                 return s;
             }
 
-            std::string wal_path;
-            s = config.GetWalConfigWalPath(wal_path);
+            bool wal_enable = false;
+            s = config.GetWalConfigEnable(wal_enable);
             if (!s.ok()) {
                 return s;
             }
 
-            if (not boost::filesystem::create_directories(wal_path)) {
-                return Status(SERVER_UNEXPECTED_ERROR, "Cannot create wal dir");
-            }
-            s = InstanceLockCheck::Check(wal_path);
-            if (!s.ok()) {
-                std::cerr << "deploy_mode: " << deploy_mode << " instance lock wal path failed." << std::endl;
-                return s;
+            if (wal_enable) {
+                std::string wal_path;
+                s = config.GetWalConfigWalPath(wal_path);
+                if (!s.ok()) {
+                    return s;
+                }
+
+                try {
+                    // True if a new directory was created, otherwise false.
+                    boost::filesystem::create_directories(wal_path);
+                } catch (...) {
+                    return Status(SERVER_UNEXPECTED_ERROR, "Cannot create wal directory");
+                }
+                s = InstanceLockCheck::Check(wal_path);
+                if (!s.ok()) {
+                    std::cerr << "deploy_mode: " << deploy_mode << " instance lock wal path failed." << std::endl;
+                    return s;
+                }
             }
         }
 
@@ -237,6 +293,13 @@ Server::Start() {
         if (!s.ok()) {
             return s;
         }
+
+#ifdef MILVUS_GPU_VERSION
+        s = GpuChecker::CheckGpuEnvironment();
+        if (!s.ok()) {
+            return s;
+        }
+#endif
         /* record config and hardware information into log */
         LogConfigInFile(config_filename_);
         LogCpuInfo();