+++ core/src/server/web_impl/component/AppComponent.hpp
@@ -11,15 +11,16 @@
 
 #pragma once
 
-#include <oatpp/parser/json/mapping/Serializer.hpp>
-#include <oatpp/parser/json/mapping/Deserializer.hpp>
-#include <oatpp/web/server/HttpRouter.hpp>
+#include <iostream>
+
+#include <oatpp/core/macro/component.hpp>
 #include <oatpp/network/client/SimpleTCPConnectionProvider.hpp>
 #include <oatpp/network/server/SimpleTCPConnectionProvider.hpp>
+#include <oatpp/parser/json/mapping/Deserializer.hpp>
+#include <oatpp/parser/json/mapping/Serializer.hpp>
 #include <oatpp/parser/json/mapping/ObjectMapper.hpp>
-#include <oatpp/core/macro/component.hpp>
-
-#include "server/web_impl/handler/WebRequestHandler.h"
+#include <oatpp/web/server/HttpConnectionHandler.hpp>
+#include <oatpp/web/server/HttpRouter.hpp>
 
 namespace milvus {
 namespace server {
@@ -41,9 +42,9 @@ class AppComponent {
         try {
             return oatpp::network::server::SimpleTCPConnectionProvider::createShared(this->port_);
         } catch (std::exception& e) {
-            std::string error_msg = "Cannot bind http port " + std::to_string(this->port_) +
-                                    ". Check if the port is already used";
-            std::cout << error_msg << std::endl;
+            std::string error_msg = "Cannot bind http port " + std::to_string(this->port_) + ". " + e.what() +
+                "(errno: " + std::to_string(errno) + ", details: " + strerror(errno) + ")";
+            std::cerr << error_msg << std::endl;
             throw std::runtime_error(error_msg);
         }
     }());
@@ -58,7 +59,7 @@ class AppComponent {
     }());
 
     OATPP_CREATE_COMPONENT(std::shared_ptr<oatpp::network::server::ConnectionHandler>, server_connection_handler_)([] {
-        OATPP_COMPONENT(std::shared_ptr<oatpp::web::server::HttpRouter>, router); // get Router component
+        OATPP_COMPONENT(std::shared_ptr<oatpp::web::server::HttpRouter>, router);
         return oatpp::web::server::HttpConnectionHandler::createShared(router);
     }());
 
@@ -66,8 +67,7 @@ class AppComponent {
         auto serializerConfig = oatpp::parser::json::mapping::Serializer::Config::createShared();
         auto deserializerConfig = oatpp::parser::json::mapping::Deserializer::Config::createShared();
         deserializerConfig->allowUnknownFields = false;
-        return oatpp::parser::json::mapping::ObjectMapper::createShared(serializerConfig,
-                                                                        deserializerConfig);
+        return oatpp::parser::json::mapping::ObjectMapper::createShared(serializerConfig, deserializerConfig);
     }());
 };